+++ tests/sqrt2.cpp
@@ -30,8 +30,8 @@ using namespace tinyopt::nlls;
 
 void TestSqrt2() {
   auto loss = [&](const auto &x, auto &grad, auto &H) {
-    double res = x * x - 2;  // since we want x to be sqrt(2), x*x should be 2
-    double J = 2 * x;        // residual's jacobian/derivative w.r.t x
+    float res = x * x - 2;  // since we want x to be sqrt(2), x*x should be 2
+    float J = 2 * x;        // residual's jacobian/derivative w.r.t x
     // Manually update the hessian and gradient
     if constexpr (!traits::is_nullptr_v<decltype(grad)>) {
       H(0, 0) = J * J;
